+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Dependency on inventory: 3.0
 * Use PropTypes, not React.PropTypes. STRIPES-427.
 * Validation. Fixes UIIN-19.
+* Show Title, Contributors, Publishers on search results. Fixes UIIN-9.
 
 ## [1.0.0](https://github.com/folio-org/ui-instances/tree/v1.0.0) (2017-09-08)
 [Full Changelog](https://github.com/folio-org/ui-items/compare/v0.0.1...v1.0.0)